Move ALL_ARCHITECTURES and imports to architecture_registry.js
Consolidates architecture ownership into a single module:
- Moves ALL_ARCHITECTURES array and ARCH_DISPLAY_NAMES from background.js
  to architecture_registry.js (single source of truth)
- Removes 80+ architecture imports from background.js (was duplicated
  with architecture_registry.js)
- background.js re-exports ALL_ARCHITECTURES and ARCH_DISPLAY_NAMES for
  existing consumers (hud.js, timeline.js, arch_selector.js, favorites.js)
- selectArchitecture() no longer needs allArchitectures parameter
- background.js import block drops from 90 lines to 16
